core & cfgparse: use grid-ratio coordinates for barrier specs (#78)
Make biosim_barrier_spec_t fully ratio-based: x, y, length and width are
now grid ratios in [0, 1] instead of integer cell coordinates, so a barrier
layout scales with the grid size. The placement functions resolve ratios to
cells (position rounded to nearest, dimensions against min(W, H)), matching
the webapp's previous boundary arithmetic so rendered output is unchanged.

- core: float x/y, BIOSIM_BARRIER_POS_UNSET sentinel becomes -1.0F, and
  ratio_to_cell/ratio_to_dim helpers in barriers.c
- cfgparse: parse x/y as floats; TOML coordinates are now ratios
- sim-wasm: biosim_wasm_add_barrier takes float x/y
- webapp: setBarriers and tomlConfig pass fractions straight through, dropping
  the fraction-to-cell conversion at both boundaries

+++ packages/sim-wasm/src/bindings.c
@@ -495,13 +495,13 @@ static int barriers_grow(void) {
     return BIOSIM_OK;
 }
 
-/* Append one barrier spec to the list. x and y are grid cell coordinates;
- * pass -32768 (INT16_MIN = BIOSIM_BARRIER_POS_UNSET) for random placement.
- * length and width are in cells; pass 0.0 (BIOSIM_BARRIER_DIM_UNSET) for a
- * random dimension. Returns BIOSIM_OK or BIOSIM_ERR_NOMEM on allocation
- * failure. */
+/* Append one barrier spec to the list. x, y, length and width are grid ratios
+ * in [0, 1]: x/y are fractions of grid width/height, length/width fractions of
+ * the smaller axis. Pass -1.0 (BIOSIM_BARRIER_POS_UNSET) for a random position
+ * or 0.0 (BIOSIM_BARRIER_DIM_UNSET) for a random dimension. Returns BIOSIM_OK
+ * or BIOSIM_ERR_NOMEM on allocation failure. */
 EMSCRIPTEN_KEEPALIVE int biosim_wasm_add_barrier(
-    int kind, int x, int y, float length, float width
+    int kind, float x, float y, float length, float width
 ) {
     int rc = barriers_grow();
     if (rc != BIOSIM_OK) {
@@ -509,8 +509,8 @@ EMSCRIPTEN_KEEPALIVE int biosim_wasm_add_barrier(
     }
     biosim_barrier_spec_t *b = &barriers[n_barriers];
     b->kind = (biosim_barrier_kind_t)kind;
-    b->x = (int16_t)x;
-    b->y = (int16_t)y;
+    b->x = x;
+    b->y = y;
     b->length = length;
     b->width = width;
     n_barriers++;
